Remove inconsistencies in handling of rating-config. Now, these values are floats in all recommenders

+++ kiwi-content/kiwi/config.py
@@ -19,7 +19,7 @@ def read_mysql_config():
 
 def read_config():
     return {
-        "max_rating": int(environ.get('MAX_RATING', 1)),
-        "min_rating": int(environ.get('MIN_RATING', 0)),
+        "max_rating": float(environ.get('MAX_RATING', 1)),
+        "min_rating": float(environ.get('MIN_RATING', 0)),
         "positive_cutoff": float(environ.get('POS_CUTOFF', 0)) # rating > cutoff -> 1, sonst -1
     }
